backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
sqlalchemy==2.0.32
alembic==1.13.3
psycopg2-binary==2.9.10
pytest==8.3.2
httpx==0.27.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyJWT>=2.8.0
